move create into user_wins controller
