+++ app/controllers/api/v1/wins_controller.rb
@@ -7,19 +7,4 @@ class Api::V1::WinsController < ApplicationController
     end
     render json: WinSerializer.new(wins)
   end
-
-  def create
-    user = User.find(params[:user_id])
-    win = user.wins.new(entry: params[:message])
-    if win.save
-      WinsChannel.broadcast_to(@wins, WinSerializer.new(win))
-      head :ok
-    end
-  end
-
-  private
-
-  def win_params
-    params.permit(:entry)
-  end
 end
