+++ config/environments/development.rb
@@ -63,7 +63,7 @@ Rails.application.configure do
   # Uncomment if you wish to allow Action Cable access from any origin.
   # config.action_cable.disable_request_forgery_protection = true
   
-  config.action_cable.url = "ws://localhost:3000/"
+  config.action_cable.url = "ws://localhost:3000/cable"
   config.action_cable.allowed_request_origins = [/http:\/\/*/, /https:\/\/*/, nil]
-  config.action_cable.mount_path="ws://localhost:3000/"
+  config.action_cable.mount_path="ws://localhost:3000/cable"
 end
